mcp-agent/pyproject.toml
[project]
name = "mcp-tool-generator"
version = "0.1.0"
description = "Automatic MCP server generator for AI4S-agent-tools"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.7.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.74.3"
]

